Convert Twist messages to a turning radius properly as opposed to treating angular.z as the current angle
The angular component of the twist depends on the linear velocity of the rover. They are related by the turning radius
That way the turning radius can be calculated properly. This changes the backwards driving behavior of the robot as
well as the steering when the robot is standing still: nothing happens when just standing still and steering. You need
to give it a slight forward or backward velocity to get it to move as the twist depends on that.